move image repo+tag configs to docker/configs.ts
