+++ src/docker/configs.ts
@@ -3,7 +3,12 @@ import { Agent, RequestInit as UndiciRequestInit } from "undici";
 import { workspace, WorkspaceConfiguration } from "vscode";
 import { SystemApi } from "../clients/docker";
 import { Logger } from "../logging";
-import { LOCAL_DOCKER_SOCKET_PATH } from "../preferences/constants";
+import {
+  LOCAL_DOCKER_SOCKET_PATH,
+  LOCAL_KAFKA_IMAGE,
+  LOCAL_KAFKA_IMAGE_TAG,
+} from "../preferences/constants";
+import { DEFAULT_KAFKA_IMAGE_REPO, DEFAULT_KAFKA_IMAGE_TAG } from "./constants";
 
 const logger = new Logger("docker.configs");
 
@@ -23,6 +28,18 @@ export function getSocketPath(): string {
   logger.debug("using docker socket path:", { socketPath: path });
 
   return path;
+}
+
+/** Get the local Kafka image name based on user settings. */
+export function getLocalKafkaImageName(): string {
+  const configs: WorkspaceConfiguration = workspace.getConfiguration();
+  return configs.get(LOCAL_KAFKA_IMAGE, DEFAULT_KAFKA_IMAGE_REPO);
+}
+
+/** Get the local Kafka image tag based on user settings. */
+export function getLocalKafkaImageTag(): string {
+  const configs: WorkspaceConfiguration = workspace.getConfiguration();
+  return configs.get(LOCAL_KAFKA_IMAGE_TAG, DEFAULT_KAFKA_IMAGE_TAG);
 }
 
 /** Default request options for Docker API requests, to be used with service class methods from `src/clients/docker/*`. */
